[DRAFT cleanup] Fail early in CreateInstanceTask if the featureName is invalid
If the feature name to which to attach the new instance is invalid,
we'll fail anyway with a FeatureNotFoundException (thrown by the
extPackage.isMany() test). But at this point we've already created an
instance for nothing, and polluted the environment with a variable
pointing to that instance (which will not be attached to the model).

Fail early if we know we will not be able to complete the task properly.

Change-Id: I6b4f9f8610705298f3d07273bf360b39ee26c23f
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>