[DRAFT 507887] Allow for negative initial table width to hide header column
Actually use the initialHeaderColumndWidth setting from the VSM to set
the initial/default width on tables. Because the default value "0"
already means "auto-size", and any positive value is taken as an
explicit width, use "-1" to mean "hide the header". For some weird
reason, the spinner widget used to set the value on the TableDescription
does not allow for negative values, so tweak the generated code to
support that.

This is just for testing, and if the runtime behavior works as expected,
some cleanups will be needed. And documentation. And tests.

Bug: 507887
Change-Id: I701776dea482c19029c6eaf204ffffc61b48d228
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>