[495036] Remove IVariableStatusListener and associated methods
This is just a cleanup, this part of the API is not used for anything
and is pure overhead. It breaks the Query Legacy interpreter, but
nobody is supposed to still use it. If someone actually complains, the
adapdation needed to make it work again is trivial. If nobody does, it
will be a clean sign that nobody cares and we can proceed with more
involved changes.

Bug: 495036
Change-Id: Ie2879abefaaed5ad42a544d85625d62f09f3934c
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>