[510040] Update CreateRepresentationWizard to use new graphic component
The wizard CreateRepresentationWizard now use the graphic component
GraphicalRepresentationHandler that has been made configurable to handle
new cases.


* The CreateRepresentationWizard has now three different usage regarding
the launch context:

** The first is to provide only the session. In this case user has to
choose a representation description and a semantic model element from
which a representation instance will be created.

 ** The second is to provide the representation description in addition
of the session. In this situation, this description will be selected by
default and the semantic page will be shown to user directly.
  
 ** The last one is to provide a semantic element in addition of the
session. In this case, the wizard will only have the page allowing to
select the representation description because the semantic element is
already provided. This page will show only representations descriptions
that are compatible with the type of the semantic element.

Bug: 510040
Change-Id: I6ca6f73ed1d828433907da03e54f65cc08f5de49
Signed-off-by: pguilet <pierre.guilet@obeo.fr>