[510040] Add enable/disable viewpoint buttons and update behavior
A button "Enable" is added to representations block allowing to activate
one or more viewpoints in the session.

A button "Disable" is added to representations block allowing to
deactivate one or more viewpoints from the session.

When deactivating a viewpoint, if some activated viewpoints depend on
it, user is asked to deactivate those as well or cancel.

When activating a viewpoint, if it misses some viewpoint dependencies,
the users is asked to activate those as well or to cancel.

Pointing at viewpoint item and representation items shows a tooltip for
better comprehension.

Activation and deactivation can only be done by selecting one or more
viewpoint. If representation descriptions or representation instances
are in the selection, an error message tells the user it cannot be done.


Bug: 510040
Change-Id: I2bd202a70022182fe4215d7963105744e8bdac4a
Signed-off-by: pguilet <pierre.guilet@obeo.fr>