[514546] Remove Guava dependency in oes.ext.gmf.runtime
We only used Preconditions.checkNotNull() in some methods before using
the parameter, but Preconditions.checkNotNull() throws an NPE if
passed null, so this is the exact same behavior with or without the
check. The check can be useful to fail early (e.g. in a setter or
constructor) instead of when the value is used at a later point or
after having made some (partial) side-effects, but these reasons do
not apply here.

Bug: 514546
Change-Id: I528193eaf3b72d539f8856a11c6c801e710ee1a7
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>